Fixed dependencies; Altered get_models.sh
