+++ Cargo.toml
@@ -7,5 +7,6 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-opencv = "0.53"
-clap = "2.33.3"
+clap = { version = "4.0.22", features = ["derive"] }
+once_cell = "1.16.0"
+opencv = "0.73.0"
